backend/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools==70.0.0"]

[project]
name = "flowey"
version = "0.1.0"
requires-python = ">=3.12.3"
dependencies = [
    "fastapi-cli-slim==0.0.4",
    "fastapi-slim==0.111.0",
    "uvicorn==0.30.1"
]

[project.optional-dependencies]
dev = [
    "build==1.2.1",
    "mypy==1.10.0",
    "ruff==0.4.9",
    "uv==0.2.11"
]

[project.scripts]
flowey = "flowey:main"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D107",
    "D203",
    "D212",
    "UP040" # Until `mypy` supports `type`
]

[tool.uv.pip]
generate-hashes = true
strict = true
